Implementation of the beginnings of the BTRFS backend.
This change implements the beginnings of the BTRFS backend.  A simple Ping()
RPC function is added into the service, as well as a client and server app
(very simple) to test the communications.

TODO: I don't currently have a good way of making the async RPC calls more
synchronous, which might be needed for Init() to call Ping().  Alternatively,
Ping() could be called as part of the negotiation protocol, and Init can just
bring up the RPC system.