Use boost's threading to make the Init() Ping synchronous.
This change adds in use of the boost futures and promises, as well as threading
to run the dispatcher in a separate thread, allowing for asynchronous RPC calls
to be made in a synchronous way.

I've also added in Google flags support, as well as turning on stack trace
dumps on fatal failures.