Implement a bit more of the protocol, move to ASIO RPC method.
We've now got an EnumerateBackupSets() RPC call with a stub implementation that
returns a lot of data.  I'm seeing lots of timeouts in the RPCs, which doesn't
bode well for transferring lots of data very quickly...It may be possible to
get around it by using lots of threads and just flooding the pipeline
asynchronously.  Then at least we'll be transferring very quickly, although we'll
definitely have to deal with errors.

Then again, this is networking, so having to deal with errors is par for the
course.  At least now the RPC system doesn't shut down completely.