Move a lot of the stuff that was in client into backend.
Much of the stuff that was in client is actually part of the backend.  The
BTRFS backend just happens to be in two parts -- a server program that runs
on a remote host, and a module that interfaces with the client application.
The client application instantiates an instance of the backend it wants.