Lots of cleanup, additional functionality.
This change separates out the client and server-side functionality
of the backup set, and implements various client-facing functions
that activate server-side RPCs to execute actions.  This allows a
BackupSet created on the client to affect server-side actions without
requiring access to the server implementation.

There's still a lot of cleanup that needs to be done -- the client and
server halves of the backend are very much intermingled and need to be
separated out.