Start making backups more useful.
This converts the Backup proto class to a client-accessible class with methods.
In doing this, I found an issue with dependencies between the client and the
backends where server-only resources were being included in client code and
vise-versa.

I've also cleaned up a bit of the ICE communcicator and adapter usage.  The
communicator and adapter created for the btrfs backend server will be the
only set ever used, so a simplifying measure is to store them in a singleton.
This way, all client/server objects can make use of it easily to create
and manage proxies.