Add a backup command to the CLI.
The CLI is now capable of initiating a backup, given a backup set ID,
backup ID, and path to a file containing filenames to back up.  The
files are sorted into buckets by their type, which will help the backend
determine what it needs to do to backup the file.  (I.e. fifos and sockets
don't need to be transferred, just re-created on the other side, and zero
files don't need to be transferred at all, but shouldn't be linked either).