Remove global.h and the IceObjects singleton.
Turns out that Ice::Current objects contain in them the ObjectAdapter,
which was the primary missing piece that prompted the creation of the
IceObjects singleton.  There were also places where the communicator
was being pulled from it where now we can either get it from the
adapter, or just pass it in on construction.