Separate out schema building from the backup implementation.
This change implements a schema class and a version 1 schema
capable of managing an SQLite database.  This should allow for
easier versioning of schemas, as the correct schema class for
each database is selected and used.

Databases are *not* upgraded to newer schemas, as that risks
corruption of the backup.  Rather, the latest version of the backup
software shall always be able to read and deal with the oldest of
backups.

BUG=#1,#7