Implement prelimiary ability to backup files.
With this change comes the ability to actually send file contents
and have them written to the BTRFS filesystem in the proper places.
There's error checking in a lot of places, but probably not all, and
handling of errors is sketchy right now.

Empty directories, special files, symlinks, and other kinds of non-
file files are not presently handled and are ignored at this point.
But, a diff against the source and target directories show the files
were 100% successfully copied over.

BUG=#1